+++ src/RekalogikaMapperBundle.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper;
 
-use Rekalogika\Mapper\DependencyInjection\CompilerPass\DebugTransformerPass;
+use Rekalogika\Mapper\DependencyInjection\CompilerPass\DebugPass;
 use Rekalogika\Mapper\DependencyInjection\CompilerPass\ObjectMapperPass;
 use Rekalogika\Mapper\DependencyInjection\CompilerPass\PropertyMapperPass;
 use Rekalogika\Mapper\DependencyInjection\CompilerPass\RemoveOptionalDefinitionPass;
@@ -37,7 +37,7 @@ class RekalogikaMapperBundle extends Bundle
         $container->addCompilerPass(new ObjectMapperPass());
 
         if ((bool) $container->getParameter('kernel.debug')) {
-            $container->addCompilerPass(new DebugTransformerPass());
+            $container->addCompilerPass(new DebugPass());
         }
     }